move getStats to router
it's a router level thing, to censor info that goes through the HTTP API
