+++ src/watcher.js
@@ -36,28 +36,6 @@ function parseLogs(interface, logs) {
     }
 }
 
-/** TODO: this might belong to state? getLatestBlockSummary, getLatestWithdrawableBlockSummary
- * @typedef {Object} BlockSummary
- * @property {Number} blockNumber
- * @property {Number} timestamp when the Monoplasma block was stored, NOT Ethereum block timestamp
- * @property {Number} memberCount
- * @property {Number} totalEarnings
- */
-
-/**
- * Don't send the full member list back, only member count
- * @returns {BlockSummary}
- */
-function summarizeBlock(block) {
-    if (!block || !block.members) { block = { members: [] } }
-    return {
-        blockNumber: block.blockNumber || 0,
-        timestamp: block.timestamp || 0,
-        memberCount: block.members.length,
-        totalEarnings: block.totalEarnings || 0,
-    }
-}
-
 /**
  * MonoplasmaWatcher hooks to the Ethereum root chain contract and Streamr join/part stream
  * It syncs the state from Ethereum and Streamr into the store
@@ -344,24 +322,5 @@ module.exports = class MonoplasmaWatcher extends EventEmitter {
     async getContractTokenBalance() {
         const balance = await this.token.methods.balanceOf(this.state.contractAddress).call()
         return balance
-    }
-
-    /**
-     * Returns the "real-time plasma" stats
-     * @returns {Object} summary of different stats and config of the community the watcher is watching
-     */
-    getStats() {
-        const joinPartStreamId = this.channel.stream.id
-        const memberCount = this.plasma.getMemberCount()
-        const totalEarnings = this.plasma.getTotalRevenue()
-        const latestBlock = summarizeBlock(this.plasma.getLatestBlock())
-        const latestWithdrawableBlock = summarizeBlock(this.plasma.getLatestWithdrawableBlock())
-        return {
-            memberCount,
-            totalEarnings,
-            latestBlock,
-            latestWithdrawableBlock,
-            joinPartStreamId,
-        }
     }
 }
